redis_store: 100-fixup — eager dispatcher init + 5 BLOCKER fixes (#100)
Reframe of the keyspace-notification dispatcher from "closes the
stale-positive class for ExistenceCacheStore wrapping RedisStore" to
**defensive infrastructure for Redis (Valkey) keyspace eviction
notifications. Currently dormant in production (Valkey on buildcache has
had zero evictions in this server's lifetime; deployed for forward-compat
against future workloads stressing the eviction path).** The original
framing overclaimed; the design doc §0 already established the
zero-eviction baseline.

Addresses 6 BLOCKER findings from the 681ae250 reviewer pass
(code-reviewer / perf-optimizer / security-reviewer / red-team /
distributed-systems-reviewer):

- BLOCKER 1 (silent init failure): convert `RedisStore::new_standard`
  to perform CONFIG GET / SET / 3× PSUBSCRIBE eagerly, before the Arc is
  handed back. Failure → `Err` from the constructor → server fails to
  start with a clear error. `register_item_callback` becomes a sync
  fast-path-only operation (no background spawn that swallows failures);
  ECS panics loudly via the `.expect()` if the dispatcher is not wired.

- BLOCKER 2 (consume-before-validate `take()`): defer
  `subscriber_channel.lock().take()` to the LAST step of eager init, so
  any failure leaves the channel still in `Mutex<Option<...>>`. The
  misleading "already consumed by SchedulerSubscriptionManager" error
  on retries is now unreachable through legitimate paths.

- BLOCKER 3 (`StoreKey::Str` over-action): `parse_keyspace_payload`
  drops non-digest payloads at the parser. Admin `DEL cas:debug-foo`
  and any non-digest scheduler/index keys never propagate to ECS or
  other ItemCallback consumers. The Str fallback was perf+log-leak
  waste, not a panic risk (`StoreKey::into_digest` blake3-hashes), but
  guaranteed-no-op cache misses + foreign-tenant info leaks are still
  worth filtering.

- BLOCKER 4 (CONFIG SET non-persistent across Valkey restart):
  `StandardRedisManager::reconnect_with` re-issues `CONFIG SET
  notify-keyspace-events <flags>` on every subscriber-slot reconnect
  before replaying PSUBSCRIBE. Without this, a Valkey restart silently
  reverts the in-memory CONFIG to whatever `valkey.conf` says (typically
  empty), the dispatcher keeps running, and ECS retains stale positives
  indefinitely. Operator should ALSO persist `notify-keyspace-events
  Eegx` in `valkey.conf` for resiliency across NativeLink restarts.

- BLOCKER 5 (shared subscriber_channel): `set_spec_defaults` rejects
  the `enable_keyspace_notifications=true` +
  `experimental_pub_sub_channel.is_some()` collision at construction
  time with `Code::FailedPrecondition`. Both consumers want the single
  Redis push-sender; the loser silently degrades to no-op invalidation
  today. Fan-out task (per design §1.1) remains future work.

- BLOCKER 6 (CI gating): wire `--features redis-integration-tests`
  into `.github/workflows/native-cargo.yaml` with a Linux-only
  `valkey-server` install step. Default `cargo test` on dev boxes
  remains a no-op for these tests (the test bin is `#![cfg(...)]`-gated
  out, not silently skipped).

Hardening (security-reviewer MEDIUMs):
- `keyspace_notifications_db: i64` → `u8` (Redis db indices are
  non-negative single bytes in practice).
- `set_spec_defaults` cross-checks `keyspace_notifications_db` against
  the URL-embedded db; mismatch is `Err` instead of silent
  listen-on-wrong-db.
- Empty `key_prefix` with `enable_keyspace_notifications=true` warns
  loudly at construction (foreign-tenant DELs would otherwise dispatch
  into our ItemCallback consumers).

Tests:
- New regression `non_digest_key_under_prefix_does_not_fire_callback`
  (BLOCKER 3): asserts a digest-shaped positive control reaches the
  callback while a same-prefix non-digest DEL is dropped at the parser.
- New regression `keyspace_and_scheduler_collision_refused_at_construction`
  (BLOCKER 5): `RedisStore::new_standard` returns `Err` when both flags
  are set; bespoke message verifies the rejection text.
- Existing `foreign_prefix_does_not_fire_callback` updated to use a
  digest-shaped positive control (the previous `cas:positive-control`
  string-key would now be filtered by BLOCKER 3 — the test relied on
  that pre-fix Str fallback).
- All 8 integration tests pass against real `valkey-server` v9.0.4
  with the `redis-integration-tests` feature enabled.

Mutation-step verification (CLAUDE.md TDD rule 5):
- BLOCKER 1: comment out `store.init_keyspace_dispatcher_eager()` in
  `new_standard` → integration test #6
  (`existence_cache_drops_positive_after_redis_eviction`) panics with
  `register_item_callback` returning Err and ECS `.expect()` firing.
- BLOCKER 3: revert `parse_keyspace_payload` to fall through to
  `StoreKey::Str(...)` → BLOCKER 3 regression FAILED as expected
  (Str leaks into captured set).
- BLOCKER 5: gate the collision check on `if false` → BLOCKER 5
  regression FAILED as expected (no Err returned from new_standard).
All three mutations reverted post-verification.


---

Squash-in (2026-05-10 — 9-reviewer cadre fix-up follow-up):
The original 7830d435 WIP commit by `parent <parent@local>` flipped
`default_enable_keyspace_notifications: true → false`. Two reviewers
(code-reviewer + DSR) caught that the flip boot-panics the production
CAS chain `cas_INNER (ExistenceCacheStore) → SizePartitioning → FastSlow
→ REDIS_CAS_SMALL_STORE`: ECS::new_with_time `.expect()`s
`register_item_callback` to succeed, FastSlow/SizePartitioning propagate
the Err, and `RedisStore::register_item_callback` returns
`Code::FailedPrecondition` whenever `enable_keyspace_notifications=false`
— putting nativelink.service into a systemd restart loop. Production
Valkey actually grants `+@all` (verified via `valkey-cli ACL LIST`), so
the original `true` default is correct.

User authorized Option 1 (revert the default flip). The bool flip is
reverted; the genuinely-useful redis_store.rs additions are folded in
here (where the eager-init they support already lives):

- `subscriber_channel` and `keyspace_dispatcher_tx` get explicit
  `UNBOUNDED-OK` justifications per the CLAUDE.md unbounded-buffer rule
  (DSR M2 + perf-optimizer M1 + code-reviewer S1).
- New `callbacks_registered_total: AtomicU64` counter, bumped on
  successful `register_item_callback`, exported via metrics.
- Inert-dispatcher 60s startup `warn!` (red-team #3 from the cadre): if
  `enable_keyspace_notifications=true` but no callback registers within
  60s, log a single warn so the operator sees the dispatcher firing into
  a void (e.g. AC-only stores not yet wrapped in ECS). Single-shot;
  uses `Arc::downgrade` so it doesn't keep the store alive.

Docstring rewritten to keep `Default: true` while tightening operator
guidance: hardened-cluster operators (no `+config` ACL) MUST set
`enable_keyspace_notifications=false` AND remove any `ExistenceCacheStore`
wrapping the `RedisStore`, or `ECS::new_with_time` panics → restart loop.

Hardened-cluster ACL operators (no `+config` grant) handled separately:
tracked as #379-pending — a config-validator that fails-fast at startup
when `enable_keyspace_notifications=true` is wrapped by ECS but the
runtime CONFIG GET fails would let operators see the misconfiguration in
their dev-loop instead of in the systemd restart loop.

See #378 for the follow-up minor-cleanup bundle from the same cadre
(security MINOR-1/2: debug-log info-leak + cfg-gating
`set_spec_defaults_for_test`; code-simplifier MINOR-1/2/3: duplicated
default-line / `.clone()` / `raw_conn` extraction; rust-crate NIT-1/2:
import ordering + safe indexing).

+++ nativelink-config/src/stores.rs
@@ -1731,19 +1731,27 @@ pub struct RedisSpec {
     /// (`__keyevent@<db>__:{del,expired,evicted}`) and dispatch them to
     /// every `ItemCallback` registered via `register_item_callback`.
     ///
-    /// When `true`, on the first `register_item_callback` call the store
-    /// issues `CONFIG SET notify-keyspace-events <merged>` (merged with any
+    /// When `true`, `RedisStore::new_standard` eagerly issues `CONFIG GET`
+    /// + `CONFIG SET notify-keyspace-events <merged>` (merged with any
     /// operator-set flags so we never trample) and `PSUBSCRIBE
-    /// __keyevent@<db>__:{del,expired,evicted}`. When `false`, those
-    /// operations are skipped and registered callbacks will never fire —
-    /// wrapper caches such as `ExistenceCacheStore` will keep stale-positive
-    /// entries when keys are silently evicted under
-    /// `maxmemory-policy=allkeys-lru`.
+    /// __keyevent@<db>__:{del,expired,evicted}` at construction. When
+    /// `false`, those operations are skipped and `register_item_callback`
+    /// returns `Err` — wrappers such as `ExistenceCacheStore` panic at
+    /// construction rather than silently retain stale-positive entries
+    /// when keys are evicted under `maxmemory-policy=allkeys-lru`.
     ///
-    /// Set to `false` if the configured Redis user lacks `CONFIG` ACLs, the
-    /// `CONFIG` command has been renamed/disabled, or notifications are
-    /// configured out-of-band and you want NativeLink to skip the runtime
-    /// mutation.
+    /// Set to `false` ONLY if the configured Redis/Valkey user lacks the
+    /// `+config` ACL (a hardened-cluster default), the `CONFIG` command
+    /// has been renamed/disabled, or notifications are configured
+    /// out-of-band and you want NativeLink to skip the runtime mutation.
+    /// In that case the operator MUST also remove any `ExistenceCacheStore`
+    /// wrapping this `RedisStore` from the config; otherwise
+    /// `ExistenceCacheStore::new_with_time` panics at construction
+    /// because `register_item_callback` returns
+    /// `Code::FailedPrecondition` (and the panic propagates up the
+    /// production CAS chain `cas_INNER → SizePartitioning → FastSlow →
+    /// REDIS_CAS_SMALL_STORE`, putting the server into a systemd restart
+    /// loop).
     ///
     /// Default: `true`
     #[serde(default = "default_enable_keyspace_notifications")]
@@ -1754,9 +1762,15 @@ pub struct RedisSpec {
     /// pattern. Has no effect in cluster mode (cluster mode keyspace
     /// notification semantics are documented as undefined in Redis).
     ///
+    /// Must agree with the database embedded in the connection URL (the path
+    /// segment of `redis://host[/db]`). NativeLink rejects mismatches at
+    /// startup so an operator who sets `redis://server/3` but leaves this at
+    /// the default `0` cannot end up listening for keyevent notifications on
+    /// the wrong database (silent stale-positive cache).
+    ///
     /// Default: 0 (the standard Redis default db).
     #[serde(default, deserialize_with = "convert_numeric_with_shellexpand")]
-    pub keyspace_notifications_db: i64,
+    pub keyspace_notifications_db: u8,
 }
 
 const fn default_enable_keyspace_notifications() -> bool {
